Separated the presentational component into multiple files, 1 per component
