+++ 12-Use-popcorn/src/presentational/Logo.js
@@ -1,0 +1,8 @@
+export default function Logo() {
+  return (
+    <div className='logo'>
+      <span role='img'>🍿</span>
+      <h1>usePopcorn</h1>
+    </div>
+  )
+}
