+++ 12-Use-popcorn/src/presentational/Main.js
@@ -1,0 +1,3 @@
+export default function Main({ children }) {
+  return <main className='main'>{children}</main>
+}
