+++ 12-Use-popcorn/src/presentational/Movie.js
@@ -1,0 +1,14 @@
+export default function Movie({ movie, onSelectMovie, onCloseMovie }) {
+  return (
+    <li onClick={() => onSelectMovie(movie.imdbID)}>
+      <img src={movie.Poster} alt={`${movie.Title} poster`} />
+      <h3>{movie.Title}</h3>
+      <div>
+        <p>
+          <span>🗓</span>
+          <span>{movie.Year}</span>
+        </p>
+      </div>
+    </li>
+  )
+}
