+++ 12-Use-popcorn/src/presentational/Navbar.js
@@ -1,0 +1,13 @@
+import Logo from './Logo'
+
+// Structural component
+function Navbar({ children }) {
+  return (
+    <nav className='nav-bar'>
+      <Logo />
+      {children}
+    </nav>
+  )
+}
+
+export default Navbar
