+++ 12-Use-popcorn/src/presentational/Search.js
@@ -1,0 +1,11 @@
+export default function Search({ query, setQuery }) {
+  return (
+    <input
+      className='search'
+      type='text'
+      placeholder='Search movies...'
+      value={query}
+      onChange={(e) => setQuery(e.target.value)}
+    />
+  )
+}
