+++ 12-Use-popcorn/src/presentational/WatchedMoviesList.js
@@ -1,0 +1,11 @@
+import WatchedMovie from './WatchedMovie'
+
+export default function WatchedMoviesList({ watched }) {
+  return (
+    <ul className='list'>
+      {watched.map((movie) => (
+        <WatchedMovie movie={movie} key={movie.imdbID} />
+      ))}
+    </ul>
+  )
+}
